cleanup duplicated component and styles. Move share-menu and tags to widgets repo
