+++ esbuild.js
@@ -1,8 +1,19 @@
 const path = require("path")
 const { globSync } = require("glob")
-const fs = require("fs-extra")
 const { sassPlugin } = require("esbuild-sass-plugin")
 const { build } = require("esbuild")
+const fs = require("fs")
+
+// copy styles
+function copyStyles() {
+  if (!fs.existsSync("dist/styles")) {
+    fs.mkdirSync("dist/styles")
+  }
+  globSync(["./src/styles/*.scss", "./src/styles/uikit/*.scss"], {
+    withFileTypes: true,
+    includeChildMatches: false
+  }).forEach(path => fs.copyFile(path.fullpath(), `./dist/styles/${path.name}`, err => err && console.log(err)))
+}
 
 const defaultConfig = {
   entryPoints: [path.resolve(__dirname, "src/index.ts"), ...globSync("src/libs/**/index.ts")],
@@ -28,3 +39,5 @@ build({
   format: "cjs",
   outdir: "dist/cjs"
 }).catch(() => process.exit(1))
+
+copyStyles()
