+++ src/styles/components/expanded-tile-swiper/base.scss
@@ -24,10 +24,6 @@ expanded-tiles {
   }
 
   .expanded-tile-wrapper {
-    .panel-overlay {
-      display: none;
-    }
-
     .panel {
       background: var(--text-tile-background);
       border-radius: var(--expanded-tile-border-radius);
