--- src/styles/templates/share-menu/share-menu.scss
@@ -1,121 +1,0 @@
-@use "../../../styles/mixins";
-
-.share-socials-popup-wrapper {
-  display: none;
-  position: absolute;
-  place-items: center;
-  height: 100%;
-  z-index: 5;
-
-  .share-socials-popup {
-    display: flex;
-    flex-direction: column;
-    width: 351px;
-    border-radius: 5px;
-    background: #fff;
-    position: relative;
-
-    .share-modal-exit {
-      position: absolute;
-      top: -12px;
-      right: -12px;
-      z-index: 2;
-    }
-  }
-
-  .ugc-inline-share-buttons {
-    align-items: baseline;
-    display: flex;
-    justify-content: space-between;
-    padding: 20px 19px;
-  }
-
-  .popup-text {
-    text-align: center;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 50%;
-    border-bottom: solid #d6d4d5 1px;
-    padding: 20px 0 12px;
-  }
-
-  .url-copy {
-    margin: 0 19px 20px;
-
-    .url-controls {
-      display: flex;
-      height: 32px;
-      align-items: flex-start;
-      border-radius: 3px;
-      border: 1px solid #d6d4d5;
-      background: #fff;
-    }
-
-    .share-url {
-      border: none;
-      outline: none;
-      background: #fff;
-      margin: 0;
-      box-shadow: none;
-      padding: 8px 12px;
-      width: 100%;
-      color: #4d4a4c;
-      font-size: 12px;
-      font-weight: 400;
-    }
-
-    .copy-button {
-      border: none;
-      border-left: 1px solid #d6d4d5;
-      outline: none;
-      background: #fff;
-      box-shadow: none;
-      cursor: pointer;
-      height: 100%;
-      padding: 9px 12px;
-      color: #1f63ca;
-      font-size: 12px;
-
-      &:hover {
-        background-color: #1f63ca;
-        color: #fff;
-      }
-    }
-
-    .copy-status {
-      display: none;
-      color: green;
-      font-weight: 500;
-      position: absolute;
-      left: 20px;
-    }
-  }
-}
-
-.share-button {
-  all: initial;
-  cursor: pointer;
-  position: absolute;
-  right: 0;
-  top: 0;
-}
-
-@include mixins.breakpoint(tablet) {
-  .share-socials-popup-wrapper {
-    right: 30%;
-    top: 30%;
-  }
-}
-
-@media only screen and (width <= 700px) {
-  .share-socials-popup-wrapper {
-    right: 20%;
-  }
-}
-
-@include mixins.breakpoint(mobile) {
-  .share-socials-popup-wrapper {
-    right: 15%;
-  }
-}
