--- src/styles/templates/tags/tags.scss
@@ -1,38 +1,0 @@
-.tile-tags {
-  bottom: 0;
-  display: flex;
-  align-items: center;
-  gap: var(--tags-gap);
-  list-style: none;
-  mask-image: linear-gradient(to left, transparent 5%, black 15%);
-  overflow-y: auto;
-  padding: 0;
-  scrollbar-width: none;
-  z-index: 2;
-
-  .tile-tag {
-    background: var(--tile-tag-background);
-    border-radius: 3px;
-    font-size: 10px;
-    font-style: normal;
-    font-weight: 400;
-    padding: 2px 4px;
-    line-height: 16px;
-    text-wrap: nowrap;
-
-    a {
-      color: var(--text-tile-tag-font-color);
-      text-decoration: none;
-    }
-
-    &.tag-inline {
-      background: var(--tile-tag-inline-background);
-      border-radius: 25px;
-      line-height: 12px;
-
-      a {
-        color: #fff;
-      }
-    }
-  }
-}
