+++ src/styles/uikit/_icons.scss
@@ -39,10 +39,10 @@
 }
 
 .icon-share {
-  width: 17px;
-  height: 17px;
+  width: 20px;
+  height: 20px;
   background-size: contain;
-  background-image: url("data:image/svg+xml,%3Csvg%20xmlns%3D%22http%3A//www.w3.org/2000/svg%22%20width%3D%2217%22%20height%3D%2217%22%20viewBox%3D%220%200%2017%2017%22%20fill%3D%22none%22%3E%3Cmask%20id%3D%22mask0_863_114055%22%20style%3D%22mask-type%3Aalpha%22%20maskUnits%3D%22userSpaceOnUse%22%20x%3D%220%22%20y%3D%220%22%20width%3D%2217%22%20height%3D%2217%22%3E%3Crect%20width%3D%2216.0161%22%20height%3D%2216.0161%22%20fill%3D%22%23D9D9D9%22/%3E%3C/mask%3E%3Cg%20mask%3D%22url%28%23mask0_863_114055%29%22%3E%3Cpath%20d%3D%22M11.3499%2015.527C10.7938%2015.527%2010.3211%2015.3226%209.93183%2014.9138C9.54255%2014.505%209.34791%2014.0087%209.34791%2013.4247C9.34791%2013.3546%209.36459%2013.1911%209.39796%2012.9342L4.7099%2010.061C4.53195%2010.2362%204.32618%2010.3735%204.09262%2010.4727C3.85905%2010.572%203.6088%2010.6216%203.34186%2010.6216C2.78574%2010.6216%202.31305%2010.4173%201.92377%2010.0085C1.53448%209.5997%201.33984%209.10332%201.33984%208.51935C1.33984%207.93538%201.53448%207.43901%201.92377%207.03023C2.31305%206.62145%202.78574%206.41706%203.34186%206.41706C3.6088%206.41706%203.85905%206.4667%204.09262%206.56597C4.32618%206.66525%204.53195%206.80248%204.7099%206.97767L9.39796%204.10454C9.37571%204.02279%209.36181%203.94395%209.35625%203.86804C9.35069%203.79212%209.34791%203.70744%209.34791%203.61401C9.34791%203.03004%209.54255%202.53367%209.93183%202.12489C10.3211%201.71611%2010.7938%201.51172%2011.3499%201.51172C11.906%201.51172%2012.3787%201.71611%2012.768%202.12489C13.1573%202.53367%2013.3519%203.03004%2013.3519%203.61401C13.3519%204.19798%2013.1573%204.69435%2012.768%205.10313C12.3787%205.51191%2011.906%205.7163%2011.3499%205.7163C11.083%205.7163%2010.8327%205.66666%2010.5992%205.56739C10.3656%205.46811%2010.1598%205.33088%209.98188%205.15569L5.29383%208.02882C5.31607%208.11057%205.32997%208.18941%205.33553%208.26532C5.3411%208.34124%205.34388%208.42592%205.34388%208.51935C5.34388%208.61279%205.3411%208.69746%205.33553%208.77338C5.32997%208.84929%205.31607%208.92813%205.29383%209.00989L9.98188%2011.883C10.1598%2011.7078%2010.3656%2011.5706%2010.5992%2011.4713C10.8327%2011.372%2011.083%2011.3224%2011.3499%2011.3224C11.906%2011.3224%2012.3787%2011.5268%2012.768%2011.9356C13.1573%2012.3444%2013.3519%2012.8407%2013.3519%2013.4247C13.3519%2014.0087%2013.1573%2014.505%2012.768%2014.9138C12.3787%2015.3226%2011.906%2015.527%2011.3499%2015.527ZM11.3499%2014.1255C11.539%2014.1255%2011.6975%2014.0583%2011.8254%2013.924C11.9533%2013.7897%2012.0173%2013.6232%2012.0173%2013.4247C12.0173%2013.2261%2011.9533%2013.0597%2011.8254%2012.9254C11.6975%2012.7911%2011.539%2012.7239%2011.3499%2012.7239C11.1608%2012.7239%2011.0024%2012.7911%2010.8744%2012.9254C10.7465%2013.0597%2010.6826%2013.2261%2010.6826%2013.4247C10.6826%2013.6232%2010.7465%2013.7897%2010.8744%2013.924C11.0024%2014.0583%2011.1608%2014.1255%2011.3499%2014.1255ZM3.34186%209.22011C3.53094%209.22011%203.68943%209.15296%203.81734%209.01865C3.94525%208.88433%204.0092%208.7179%204.0092%208.51935C4.0092%208.3208%203.94525%208.15437%203.81734%208.02006C3.68943%207.88575%203.53094%207.81859%203.34186%207.81859C3.15278%207.81859%202.99429%207.88575%202.86638%208.02006C2.73847%208.15437%202.67452%208.3208%202.67452%208.51935C2.67452%208.7179%202.73847%208.88433%202.86638%209.01865C2.99429%209.15296%203.15278%209.22011%203.34186%209.22011ZM11.3499%204.31477C11.539%204.31477%2011.6975%204.24762%2011.8254%204.1133C11.9533%203.97899%2012.0173%203.81256%2012.0173%203.61401C12.0173%203.41546%2011.9533%203.24903%2011.8254%203.11471C11.6975%202.9804%2011.539%202.91325%2011.3499%202.91325C11.1608%202.91325%2011.0024%202.9804%2010.8744%203.11471C10.7465%203.24903%2010.6826%203.41546%2010.6826%203.61401C10.6826%203.81256%2010.7465%203.97899%2010.8744%204.1133C11.0024%204.24762%2011.1608%204.31477%2011.3499%204.31477Z%22%20fill%3D%22%23908E8F%22/%3E%3C/g%3E%3C/svg%3E");
+  background-image: url("data:image/svg+xml,%3Csvg width='19' height='20' viewBox='0 0 19 20' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cg filter='url(%23filter0_d_977_183640)'%3E%3Cpath d='M12.9167 15.8346C12.4306 15.8346 12.0174 15.6645 11.6772 15.3242C11.3369 14.9839 11.1667 14.5707 11.1667 14.0846C11.1667 14.0263 11.1813 13.8902 11.2105 13.6763L7.11258 11.2846C6.95703 11.4305 6.77716 11.5447 6.573 11.6273C6.36883 11.71 6.15008 11.7513 5.91675 11.7513C5.43064 11.7513 5.01744 11.5812 4.67716 11.2409C4.33689 10.9006 4.16675 10.4874 4.16675 10.0013C4.16675 9.51519 4.33689 9.102 4.67716 8.76172C5.01744 8.42144 5.43064 8.2513 5.91675 8.2513C6.15008 8.2513 6.36883 8.29262 6.573 8.37526C6.77716 8.4579 6.95703 8.57214 7.11258 8.71797L11.2105 6.3263C11.1911 6.25825 11.1789 6.19262 11.174 6.12943C11.1692 6.06623 11.1667 5.99575 11.1667 5.91797C11.1667 5.43186 11.3369 5.01866 11.6772 4.67839C12.0174 4.33811 12.4306 4.16797 12.9167 4.16797C13.4029 4.16797 13.8161 4.33811 14.1563 4.67839C14.4966 5.01866 14.6667 5.43186 14.6667 5.91797C14.6667 6.40408 14.4966 6.81727 14.1563 7.15755C13.8161 7.49783 13.4029 7.66797 12.9167 7.66797C12.6834 7.66797 12.4647 7.62665 12.2605 7.54401C12.0563 7.46137 11.8765 7.34714 11.7209 7.2013L7.623 9.59297C7.64244 9.66102 7.6546 9.72665 7.65946 9.78984C7.66432 9.85304 7.66675 9.92352 7.66675 10.0013C7.66675 10.0791 7.66432 10.1496 7.65946 10.2128C7.6546 10.276 7.64244 10.3416 7.623 10.4096L11.7209 12.8013C11.8765 12.6555 12.0563 12.5412 12.2605 12.4586C12.4647 12.376 12.6834 12.3346 12.9167 12.3346C13.4029 12.3346 13.8161 12.5048 14.1563 12.8451C14.4966 13.1853 14.6667 13.5985 14.6667 14.0846C14.6667 14.5707 14.4966 14.9839 14.1563 15.3242C13.8161 15.6645 13.4029 15.8346 12.9167 15.8346ZM12.9167 14.668C13.082 14.668 13.2206 14.6121 13.3324 14.5003C13.4442 14.3885 13.5001 14.2499 13.5001 14.0846C13.5001 13.9194 13.4442 13.7808 13.3324 13.669C13.2206 13.5572 13.082 13.5013 12.9167 13.5013C12.7515 13.5013 12.6129 13.5572 12.5011 13.669C12.3893 13.7808 12.3334 13.9194 12.3334 14.0846C12.3334 14.2499 12.3893 14.3885 12.5011 14.5003C12.6129 14.6121 12.7515 14.668 12.9167 14.668ZM5.91675 10.5846C6.08203 10.5846 6.22057 10.5287 6.33237 10.4169C6.44418 10.3051 6.50008 10.1666 6.50008 10.0013C6.50008 9.83602 6.44418 9.69748 6.33237 9.58568C6.22057 9.47387 6.08203 9.41797 5.91675 9.41797C5.75147 9.41797 5.61293 9.47387 5.50112 9.58568C5.38932 9.69748 5.33341 9.83602 5.33341 10.0013C5.33341 10.1666 5.38932 10.3051 5.50112 10.4169C5.61293 10.5287 5.75147 10.5846 5.91675 10.5846ZM12.9167 6.5013C13.082 6.5013 13.2206 6.4454 13.3324 6.33359C13.4442 6.22179 13.5001 6.08325 13.5001 5.91797C13.5001 5.75269 13.4442 5.61415 13.3324 5.50234C13.2206 5.39054 13.082 5.33464 12.9167 5.33464C12.7515 5.33464 12.6129 5.39054 12.5011 5.50234C12.3893 5.61415 12.3334 5.75269 12.3334 5.91797C12.3334 6.08325 12.3893 6.22179 12.5011 6.33359C12.6129 6.4454 12.7515 6.5013 12.9167 6.5013Z' fill='white'/%3E%3C/g%3E%3Cdefs%3E%3Cfilter id='filter0_d_977_183640' x='0.166748' y='0.167969' width='18.5' height='19.668' filterUnits='userSpaceOnUse' color-interpolation-filters='sRGB'%3E%3CfeFlood flood-opacity='0' result='BackgroundImageFix'/%3E%3CfeColorMatrix in='SourceAlpha' type='matrix' values='0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0' result='hardAlpha'/%3E%3CfeOffset/%3E%3CfeGaussianBlur stdDeviation='2'/%3E%3CfeComposite in2='hardAlpha' operator='out'/%3E%3CfeColorMatrix type='matrix' values='0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0'/%3E%3CfeBlend mode='normal' in2='BackgroundImageFix' result='effect1_dropShadow_977_183640'/%3E%3CfeBlend mode='normal' in='SourceGraphic' in2='effect1_dropShadow_977_183640' result='shape'/%3E%3C/filter%3E%3C/defs%3E%3C/svg%3E%0A");
 }
 
 .icon-instagram-share {
